24/03/13 커밋 변경사항 구현중, ADD 테이블
