+++ datahub_back/src/main/kotlin/com/example/datahub_back/controller/treeController/CommitController.kt
@@ -1,7 +1,7 @@
 package com.example.datahub_back.controller.treeController
 
 import com.example.datahub_back.service.treeService.*
-import com.example.datahub_back.service.treeService.commit.TestCommitService
+import com.example.datahub_back.service.treeService.commit.CommitService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PostMapping
@@ -13,12 +13,12 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping("/api/commit")
 class CommitController(
 //    private val commitService: CommitService,
-    private val testCommitService: TestCommitService
+    private val commitService: CommitService
 ) {
     @PostMapping("/")
     fun handleCommit(@RequestBody commitDataList: CommitRequest): ResponseEntity<String> {
         return try {
-            testCommitService.handleCommit(commitDataList)
+            commitService.handleCommit(commitDataList)
             ResponseEntity.ok("Commit processed successfully")
         } catch (e: Exception) {
             ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error occurred: ${e.message}")
