+++ datahub_back/src/main/kotlin/com/example/datahub_back/data/treeData/ChangePageData.kt
@@ -5,13 +5,15 @@ import com.example.datahub_back.dto.treeDTO.ChangePage
 val changePage1 = ChangePage(
     pageId = 1,
     pageName = "페이지1",
-    path = "C:\\ProgramData"
+    path = "C:\\ProgramData",
+    commit = commit1
 )
 
 val changePage2 = ChangePage(
     pageId = 2,
     pageName = "페이지2",
-    path = "C:\\ProgramData"
+    path = "C:\\ProgramData",
+    commit = commit1
 )
 
 val changePageList = mutableListOf(changePage1, changePage2)
