+++ datahub_back/src/main/kotlin/com/example/datahub_back/data/treeData/ColumnData.kt
@@ -4,8 +4,9 @@ import com.example.datahub_back.dto.treeDTO.SourceColumn
 
 val sourceColumn1 = SourceColumn(
     columnId = 1,
-    tableId = 1,
+    table = table1,
     columnName = "Column 1",
+    comment = "comment1",
     dataType = "String",
     isPrimaryKey = 1,
     isForeignKey = 0,
@@ -15,8 +16,9 @@ val sourceColumn1 = SourceColumn(
 
 val sourceColumn2 = SourceColumn(
     columnId = 2,
-    tableId = 1,
+    table = table1,
     columnName = "Column 2",
+    comment = "comment1",
     dataType = "Int",
     isPrimaryKey = 0,
     isForeignKey = 0,
@@ -26,8 +28,9 @@ val sourceColumn2 = SourceColumn(
 
 val sourceColumn3 = SourceColumn(
     columnId = 3,
-    tableId = 1,
+    table = table1,
     columnName = "Column 3",
+    comment = "comment1",
     dataType = "Boolean",
     isPrimaryKey = 0,
     isForeignKey = 0,
