+++ datahub_back/src/main/kotlin/com/example/datahub_back/dto/treeDTO/TreeDTO.kt
@@ -21,9 +21,6 @@ data class Commit(
     val comment: String,
     val createTime: LocalDateTime,
     val createUser: String,
-    val sourceTables: List<SourceTable>?,
-    val changeTables: List<ChangeTable>?,
-    val changePageIds: List<Long>?
 )
 
 data class SourceTable(
@@ -32,13 +29,15 @@ data class SourceTable(
     val comment: String,
     val isFavorite: Int,
     val isDelete: Int,
-    val updateTime: LocalDateTime
+    val updateTime: LocalDateTime,
+    val commit: Commit,
 )
 
 data class SourceColumn(
-    val tableId: Long, // FK
     val columnId: Long, // PK
+    val table: SourceTable, // PK, FK
     val columnName: String,
+    val comment : String,
     val dataType: String,
     val isPrimaryKey: Int,
     val isForeignKey: Int,
@@ -48,25 +47,40 @@ data class SourceColumn(
 
 data class SourceData(
     val dataId: Long, // PK
-    val columnId: Long, // FK
+    val column: SourceColumn, // PK, FK
     val columnLine : Int,
     val data: String
 )
 
 // 테이블 변경사항
 data class ChangeTable(
-    val changeTableId: Long, // PK, FK Table의 tableId와 일치해야 함
-    val changeTableName: String,
+    var tableId: Long, // PK
+    val tableName: String,
+    val comment: String,
+    val isFavorite: Int,
+    val isDelete: Int,
+    val updateTime: LocalDateTime,
+    val commit: Commit,
     val action: ChangeAction, // 변경 액션 (추가, 삭제)
-    val primaryKey: String?, // 변경된 행의 기본 키 이름
-    val columns: MutableList<ChangeTableColumn> // 변경된 컬럼들과 그 값들
 )
 
-data class ChangeTableColumn(
-    val changeTableId: Long, //PK, FK
-    val columnIndex: Int,
+data class ChangeColumn(
+    val columnId: Long, // PK
+    val table: ChangeTable, // PK, FK
     val columnName: String,
-    val columnValue: String
+    val comment : String,
+    val dataType: String,
+    val isPrimaryKey: Int,
+    val isForeignKey: Int,
+    val isUniqueKey: Int,
+    val joinSourceTable: SourceTable?
+)
+
+data class ChangeData(
+    val dataId: Long, // PK
+    val column: ChangeColumn, // PK, FK
+    val columnLine : Int,
+    val data: String
 )
 
 enum class ChangeAction {
@@ -78,5 +92,6 @@ enum class ChangeAction {
 data class ChangePage(
     val pageId: Long, // PK, Page의 pageId와 일치해야 함
     val pageName: String,
-    val path: String
+    val path: String,
+    val commit: Commit,
 )
