+++ datahub_back/src/main/kotlin/com/example/datahub_back/service/treeService/commit/AddChangesAsData.kt
@@ -1,0 +1,4 @@
+package com.example.datahub_back.service.treeService.commit
+
+class AddChangesAsData {
+}
