+++ datahub_back/src/main/kotlin/com/example/datahub_back/service/treeService/commit/AddChangesAsTable.kt
@@ -1,0 +1,8 @@
+package com.example.datahub_back.service.treeService.commit
+
+import org.springframework.stereotype.Service
+
+@Service
+class AddChangesAsTable {
+
+}
